+++ matchbox-frontend/src/app/util/operation-result.ts
@@ -36,8 +36,10 @@ export class Issue {
   line?: number;
   col?: number;
   sliceInfo: string[] = [];
+  markdown: boolean;
+  details: string;
 
-  constructor(severity: IssueSeverity, code: string, text: string, expression?: string, line?: number, col?: number, sliceInfo?: string[]) {
+  constructor(severity: IssueSeverity, code: string, text: string, expression?: string, line?: number, col?: number, sliceInfo?: string[], markdown?: boolean, details?: string) {
     this.severity = severity;
     this.code = code;
     this.text = text;
@@ -45,6 +47,11 @@ export class Issue {
     this.line = line;
     this.col = col;
     this.sliceInfo = sliceInfo ?? [];
+    this.markdown = markdown;
+    this.details = details;
+    if (markdown && this.text.includes("```markdown")) {
+      this.text = this.text.substring(11, this.text.length - 3);
+    }
   }
 
   static fromOoIssue(ooIssue: fhir.r4.OperationOutcomeIssue): Issue {
@@ -64,6 +71,8 @@ export class Issue {
     } else {
       text = ooIssue.diagnostics;
     }
+    let markdown = Issue.getExtensionStringValue(ooIssue, "http://hl7.org/fhir/StructureDefinition/rendering-style") == "markdown";
+    let details = ooIssue.details ? ooIssue.details.text : undefined;
     return new Issue(
       ooIssue.severity as IssueSeverity,
       ooIssue.code,
@@ -71,7 +80,9 @@ export class Issue {
       expression,
       Issue.getLineNo(ooIssue),
       Issue.getColNo(ooIssue),
-      sliceInfo
+      sliceInfo,
+      markdown,
+      details
     );
   }
 
@@ -95,6 +106,18 @@ export class Issue {
     for (const ext of issue.extension) {
       if (ext.url === url) {
         return ext.valueInteger;
+      }
+    }
+    return undefined;
+  }
+
+  static getExtensionStringValue(issue: fhir.r4.OperationOutcomeIssue, url: string): string | undefined {
+    if (!issue.extension) {
+      return undefined;
+    }
+    for (const ext of issue.extension) {
+      if (ext.url === url) {
+        return ext.valueString;
       }
     }
     return undefined;